RR:  - spelling corrections  - added todos for missing documentation  - some dummy implementation for ODE monitoring service  - changed BigInteger to int  - added some XXX's to model's system commands since they are currently unusable to the ODE monitor

+++ src/ODEMonitorClient/trunk/src/main/java/de/decidr/odemonitor/service/GetConfigResponse.java
@@ -16,7 +16,6 @@
 
 package de.decidr.odemonitor.service;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
@@ -38,7 +37,7 @@ import javax.xml.datatype.XMLGregorianCalendar;
  *   &lt;complexContent&gt;
  *     &lt;restriction base=&quot;{http://www.w3.org/2001/XMLSchema}anyType&quot;&gt;
  *       &lt;sequence&gt;
- *         &lt;element name=&quot;updateInterval&quot; type=&quot;{http://www.w3.org/2001/XMLSchema}integer&quot;/&gt;
+ *         &lt;element name=&quot;updateInterval&quot; type=&quot;{http://www.w3.org/2001/XMLSchema}int&quot;/&gt;
  *         &lt;element name=&quot;configChanged&quot; type=&quot;{http://www.w3.org/2001/XMLSchema}dateTime&quot;/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
@@ -54,7 +53,7 @@ import javax.xml.datatype.XMLGregorianCalendar;
 public class GetConfigResponse {
 
     @XmlElement(required = true)
-    protected BigInteger updateInterval;
+    protected Integer updateInterval;
     @XmlElement(required = true)
     @XmlSchemaType(name = "dateTime")
     protected XMLGregorianCalendar configChanged;
@@ -62,10 +61,10 @@ public class GetConfigResponse {
     /**
      * Gets the value of the updateInterval property.
      * 
-     * @return possible object is {@link BigInteger }
+     * @return possible object is {@link Integer }
      * 
      */
-    public BigInteger getUpdateInterval() {
+    public Integer getUpdateInterval() {
         return updateInterval;
     }
 
@@ -73,10 +72,10 @@ public class GetConfigResponse {
      * Sets the value of the updateInterval property.
      * 
      * @param value
-     *            allowed object is {@link BigInteger }
+     *            allowed object is {@link Integer }
      * 
      */
-    public void setUpdateInterval(BigInteger value) {
+    public void setUpdateInterval(Integer value) {
         this.updateInterval = value;
     }
